feat: 스터디 탈퇴 리팩토링
